+++ shared/blockchainDBInterface.go
@@ -12,7 +12,7 @@ import (
 // It includes methods for managing balances, transactions, blocks, and public keys.
 
 type BlockchainDBInterface interface {
-	GetBalance(address string, utxos map[string]UTXO) (int, error)
+	GetBalance(address string, utxos map[string]UTXO) (int64, error)
 	SendTransaction(fromAddress, toAddress string, amount int, privKey *rsa.PrivateKey) (bool, error)
 	SanitizeAndFormatAddress(address string) (string, error)
 	InsertBlock(data []byte, blockNumber int) error
@@ -20,7 +20,7 @@ type BlockchainDBInterface interface {
 	RetrievePublicKeyFromAddress(address string) (ed25519.PublicKey, error)
 	AddTransaction(tx *thrylos.Transaction) error
 	UpdateUTXOs(inputs []UTXO, outputs []UTXO) error
-	CreateUTXO(id, txID string, index int, address string, amount int) (UTXO, error)
+	CreateUTXO(id, txID string, index int, address string, amount int64) (UTXO, error)
 	GetUTXOsForUser(address string, utxos map[string]UTXO) ([]UTXO, error)
 	GetUTXOs(address string) (map[string][]UTXO, error)
 	CreateAndSignTransaction(txID string, inputs, outputs []UTXO, privKey *rsa.PrivateKey) (Transaction, error)
